Time: 0 ms (100%), Space: 21.8 MB (76.24%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -12,7 +12,7 @@
 class Solution {
 public:
     bool isValidBST(TreeNode* root) {
-        return isValidBST(root, (long long) LLONG_MIN , (long long) LLONG_MAX);
+        return isValidBST(root, LLONG_MIN , LLONG_MAX);
     }
 private:
     bool isValidBST(TreeNode* root, long long leftBound, long long rightBound) {
